+++ src/main/java/com/jkoolcloud/jesl/tnt4j/sink/JKCloudEventSinkFactory.java
@@ -49,8 +49,7 @@ public class JKCloudEventSinkFactory extends LoggedEventSinkFactory {
 	private String url = System.getProperty("jesl.sink.factory.socket.url", "http://localhost:6580");
 	// NOTE: server side uses 5min. to close inactive connection by default
 	private long idleTimeout = Long.getLong("jesl.sink.factory.socket.idle.timeout", TimeUnit.MINUTES.toMillis(4));
-	private int connTimeout = Integer.getInteger("jesl.sink.factory.socket.conn.timeout",
-			(int) HttpClient.DEFAULT_CON_TIMEOUT);
+	private long connTimeout = Integer.getInteger("jesl.sink.factory.socket.conn.timeout", HttpClient.DEFAULT_CON_TIMEOUT);
 	private boolean ackSends = Boolean.getBoolean("jesl.sink.factory.ack.sends");
 
 	private String proxyScheme = "http";
@@ -106,7 +105,7 @@ public class JKCloudEventSinkFactory extends LoggedEventSinkFactory {
 
 		url = Utils.getString("Url", settings, url);
 		token = Utils.getString("Token", settings, token);
-		connTimeout = Utils.getInt("ConnTimeout", settings, connTimeout);
+		connTimeout = Utils.getLong("ConnTimeout", settings, connTimeout);
 		idleTimeout = Utils.getLong("IdleTimeout", settings, idleTimeout);
 		proxyScheme = Utils.getString("ProxyScheme", settings, proxyScheme);
 		proxyHost = Utils.getString("ProxyHost", settings, proxyHost);
